requirements.txt
mxnet-cu100==1.5.1
py-cpuinfo==8.0.0
GPUtil==1.4.0
psutil==5.9.0
torch==1.10.1
torchvision==0.11.2
pytz==2021.3
opencv-python==4.5.4.60
insightface==0.1.5